Adding modification to run with VS2022, .net core 6.0 and Mono3.8

+++ BouncingBox/Game1.cs
@@ -2,10 +2,10 @@ using System;
 using System.Collections.Generic;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Content;
-using Microsoft.Xna.Framework.GamerServices;
+//using Microsoft.Xna.Framework.GamerServices;
 using Microsoft.Xna.Framework.Graphics;
 using Microsoft.Xna.Framework.Input;
-using Microsoft.Xna.Framework.Storage;
+//using Microsoft.Xna.Framework.Storage;
 
 using Microsoft.Xna.Framework.Audio;
 using Microsoft.Xna.Framework.Media;
@@ -40,7 +40,7 @@ namespace Microsoft.Xna.Samples.BouncingBox
 
 		private void GetNewColor ()
 		{
-			backColor = new Color ((byte)randomizer.Next (255),(byte)randomizer.Next (255),(byte)randomizer.Next (255),255);
+			backColor = new Color ((byte)randomizer.Next (255),(byte)randomizer.Next (255),(byte)randomizer.Next (255),(byte)255);
 		}
 
 		/// <summary>
